Feat: 식사권 취소 API, 포인트 타입 Long => BigDecimal TicketController : 식사권 취소하는 컨트롤러 추가, swagger 적용, 낙찰자와 주최자 전부 이 컨트롤러를 통해 취소 UserProfileDto : 포인트 타입이 BigDecimal로 바뀌면서 이 부분의 포인트 타입도 BigDecimal로 변경 DetailTicketResDto : 식사권 상세 페이지에서 기부 금액을 보여줄 때 double을 이용한 연산이 있어서 donation 타입도 BigDecimal로 변경 Ticket : ticketIsOpened 필드 값을 변경해주기 위한 메서드 생성 User : point 타입을 BigDecimal로 변경하면서 point 관련 메서드들도 알맞게 변경 ErrorCode : 식사권 취소에 필요한 에러 코드 추가 AuctionService : point 타입 변환으로 인해 에러 발생하는 부분 알맞게 변경 PaymentService : point 타입 변환으로 인해 에러 발생하는 부분 알맞게 변경 SchedulerService : point 타입 변환으로 인해 에러 발생하는 부분 알맞게 변경 TicketService : 낙찰자와 주최자가 식사권 취소하는 로직 추가

+++ src/main/java/shootingstar/var/Service/SchedulerService.java
@@ -1,5 +1,6 @@
 package shootingstar.var.Service;
 
+import java.math.BigDecimal;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -45,7 +46,7 @@ public class SchedulerService {
             User user = userRepository.findById(userId)
                     .orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
             log.info("사용자 증가 전 포인트 : {}", user.getPoint());
-            user.increasePoint(auction.getMinBidAmount());
+            user.increasePoint(BigDecimal.valueOf(auction.getMinBidAmount()));
             log.info("사용자 증가 후 포인트 : {}", user.getPoint());
 
             ScheduledTask task = scheduledTaskRepository.findById(scheduledTaskId)
